requirements.txt
numpy==1.19.5
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
pandas==1.2.1
psycopg2==2.8.6
requests==2.25.1
ShopifyAPI==8.2.0
